+++ src/Components/Footer/Footer.scss
@@ -1,0 +1,5 @@
+#Footer {
+  width: 100%;
+  height: 10vh;
+  background-color: red;
+}
