+++ src/Pages/Main/Components/BackgroundLine.js
@@ -1,0 +1,18 @@
+import React from "react";
+import "./BackgroundLine.scss";
+
+class BackgroundLine extends React.Component {
+  render() {
+    return (
+      <div class="BackgroundLine">
+        <span className="grayBackground" />
+        <span className="grayBackground" />
+        <span className="grayBackground" />
+        <span />
+        <span />
+      </div>
+    );
+  }
+}
+
+export default BackgroundLine;
