+++ src/Pages/Main/Components/ConcertContents.js
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+import poster from "./Images/post_3.jpg";
+import "./ConcertContents.scss";
+
+class ConcertContents extends Component {
+  render() {
+    return (
+      <div>
+        <img src={poster} alt="포스터" />
+        <p>쇼팽으로 만나는 겨울나라</p>
+      </div>
+    );
+  }
+}
+export default ConcertContents;
