+++ src/Pages/Main/Components/ConcertContents.scss
@@ -1,0 +1,12 @@
+.ConcertContents {
+  img {
+    width: 287px;
+    height: 405px;
+  }
+  p {
+    width: 287px;
+    height: 20px;
+    margin: 20px 0;
+    font-size: 18px;
+  }
+}
