+++ src/Pages/Main/Components/ConcertSection.scss
@@ -1,17 +1,6 @@
 @import "../../../styles/common.scss";
 
-.backgroundLine {
-  width: 100%;
-  height: 88vh;
-  position: absolute;
-
-  span {
-    width: 16.7%;
-    height: 88%;
-    float: left;
-    opacity: 1;
-    border-right: 1px solid #ececec;
-  }
+.BackgroundLine {
   .backgroundTxt {
     width: 100%;
     position: absolute;
@@ -71,7 +60,6 @@
         font-size: 24px;
         font-weight: 300;
         color: #767676;
-
         animation-name: introduce;
         animation-duration: 1.5s;
         animation-duration: leaner;
@@ -151,13 +139,4 @@
       }
     }
   }
-}
-
-#upcomingConcertBox {
-  animation-name: jaehee;
-  animation-duration: 2s;
-  animation-duration: leaner;
-  animation-iteration-count: 1;
-  animation-direction: alternate;
-  animation-fill-mode: forwards;
 }
