+++ src/Pages/Main/Components/ConcertSlideBar.scss
@@ -4,18 +4,7 @@
   position: absolute;
   top: 20vh;
   right: -20vh;
-  div {
-    img {
-      width: 287px;
-      height: 405px;
-    }
-    p {
-      width: 287px;
-      height: 20px;
-      margin: 20px 0;
-      font-size: 18px;
-    }
-  }
+
   .swiper-scrollbar {
     height: 2px;
     width: 1px;
