--- src/Pages/Main/Components/Footer.js
@@ -1,13 +1,0 @@
-import React from "react";
-import "./Footer.scss";
-
-class Footer extends React.Component {
-  render() {
-    return (
-      <div id="Footer" className="0">
-        Footertest
-      </div>
-    );
-  }
-}
-export default Footer;
