--- src/Pages/Main/Components/Footer.scss
@@ -1,5 +1,0 @@
-#Footer {
-  width: 100%;
-  height: 10vh;
-  background-color: red;
-}
