+++ src/Pages/Main/Main.scss
@@ -16,3 +16,7 @@ body {
 :root {
   --swiper-theme-color: #4a5663;
 }
+
+button {
+  outline: none;
+}
